bot/requirements.txt
ansimarkup==1.5.0
pyfiglet==0.8.post1
python-telegram-bot==13.10
PyYAML==6.0
